+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to [bpmnlint-plugin-camunda-compat](https://github.com/camun
 
 ___Note:__ Yet to be released changes appear here._
 
+* `FEAT`: validate `bpmn:AdHocSubProcess` implementation ([#214](https://github.com/camunda/bpmnlint-plugin-camunda-compat/pull/214))
 * `FEAT`: validate `outputCollection` and `outputElement` on `zeebe:adHoc`
 
 ## 2.39.1